+++ src/sph_geom/elliptical_cone.rs
@@ -37,17 +37,25 @@ impl EllipticalCone {
       theta_sin_cos,
     }
   }
-  
+
   /// Returns `true` if the given point on the unit sphere is inside the elliptical cone.
   /// # Inputs
   /// - `lon` longitude of the point we want to test, in radians
   /// - `lat` latitude of the point we want to test, in radians
-  pub fn contains(&self, lon: f64, lat: f64)-> bool {
+  pub fn contains(&self, lon: f64, lat: f64) -> bool {
     match self.center.proj(lon, lat) {
       Some((x, y)) => self.ellipse.contains(x, y),
       None => false,
     }
   }
+
+  /*pub fn may_overlap_cone(&self, lon: f64, lat: f64, radius: f64) -> bool {// !! TEST DISTANCE < r+R
+    if self.a + radius > HALF_PI { // in this case, possible out-of-bound projection for the center of the cone
+      true
+    } else {
+      self.overlap_cone(lon, lat, radius)
+    }
+  }*/
   
   /// Returns `true` if the given cone overlap the elliptical cone.
   /// # Inputs
@@ -55,13 +63,47 @@ impl EllipticalCone {
   /// - `lat` latitude of the center of the cone, in radians
   /// - `radius` cone radius, in radians
   pub fn overlap_cone(&self, lon: f64, lat: f64, radius: f64) -> bool {
-    match self.center.proj(lon, lat) {
-      Some((x, y)) => {
-        let eucl_a = (self.a + radius).sin();
-        let eucl_b = (self.b + radius).sin();
-        Ellipse::from_oriented(eucl_a, eucl_b, self.theta_sin_cos).contains(x, y)
-      },
-      None => false,
+    if self.a + radius <= HALF_PI {
+      match self.center.proj(lon, lat) {
+        Some((x, y)) => {
+          let eucl_a = (self.a + radius).sin();
+          let eucl_b = (self.b + radius).sin();
+          Ellipse::from_oriented(eucl_a, eucl_b, self.theta_sin_cos).contains(x, y)
+        },
+        None => false,
+      }
+    } else {
+      let ((x, y), ang_dist, top_hemisphere) =  self.center.forced_proj_and_distance(lon, lat);
+      // The projection of a cone on a plane is an ellipse
+      let proj_d_min = (ang_dist - radius).sin();
+      let proj_d_max = (ang_dist + radius).sin();
+      let proj_a = radius.sin();                     // projected ellipse semi-major axis
+      let proj_b = 0.5 * (proj_d_max - proj_d_min).abs(); // projected ellipse semi-minor axis
+      // ellipse theta angle = tan(t + pi/2) = cos(t) / (-sin(t)) = x / -y
+      // let angle = (x / -y).atan(); //
+      let norm = (x.pow2() + y.pow2()).sqrt(); // distance from (0.0) the cone projected center (!= ellipse center)
+      // phi = angle of the (x, y) position in the euclidean plane
+      let cos_phi = x / norm;
+      let sin_phi = y / norm;
+      let sin_cos_angle = if x >= 0.0 {
+        (cos_phi, -sin_phi)
+      } else {
+        (-cos_phi, sin_phi)
+      };
+      let proj_ell = Ellipse::from_oriented(proj_a, proj_b, sin_cos_angle);
+      let proj_ell_dist = 0.5 * (proj_d_max + proj_d_min); // distance from (0, 0) to ellipse center
+      let proj_ell_x = proj_ell_dist * cos_phi;
+      let proj_ell_y = proj_ell_dist * sin_phi;
+      if top_hemisphere {
+        self.ellipse.overlap(proj_ell_x, proj_ell_y, &proj_ell)
+      } else {
+        // The projection of the center of the cone is in the back hemisphere
+        // so the area to be considered look like a crescent moon.
+        // I am not sure that the following is OK, let's try...
+        // So far we apply the same algo, knowing that at deeper depth cells will have less changes
+        // to spuriously overalp
+        self.ellipse.overlap(proj_ell_x, proj_ell_y, &proj_ell)
+      }
     }
   }
 
@@ -90,5 +132,5 @@ impl EllipticalCone {
     }
     ar
   }
-  
+
 }